Prototype #2 New Map, 10 Cars with IDs
Created prefab of car, assigned names to the cars which operate as IDs,
all created in game setup. The instances of the car gameobject poll
between 2 points repeatedly and instantiate on a random X coordinate on
the 0y coordinate. Also implemented a new  printed map of a section of
New York. Source: Google Maps.